# backend/core/policy_engine/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grandway Policy-Registered API",
    "version": "1.0.0",
    "description": "Generated from the Core Policy Engine registry (source of truth for endpoints, permission keys, HTTP methods, and risk levels). Request/response body schemas are documented per app in `docs/DATA_CONTRACT.md`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/applicants/{applicant_id}/status/": {
      "post": {
        "operationId": "applicants.applicant.change_status",
        "summary": "Change Applicant Status",
        "description": "Set the applicant's standing: active, dormant, or archived. Always manual.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicants.applicant.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "applicant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/auth/users/{user_id}/restore/": {
      "post": {
        "operationId": "authenticate.user.restore",
        "summary": "Restore Account",
        "description": "Reactivate a blocked managed account.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "authenticate"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "authenticate.user.restore",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/checklists/{checklist_id}/archive/": {
      "post": {
        "operationId": "checklists.checklist.archive",
        "summary": "Archive Checklist",
        "description": "Take a checklist out of active work. A reason is required. An archived checklist no longer blocks re-applying its template, so this is also how staff ask for a fresh copy.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "checklists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "checklists.checklist.archive",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/checklists/{checklist_id}/restore/": {
      "post": {
        "operationId": "checklists.checklist.restore",
        "summary": "Restore Checklist",
        "description": "Return an archived checklist to work — active if it was ever activated, draft otherwise.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "checklists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "checklists.checklist.restore",
        "x-risk-level": "low",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "checklist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/clients/{client_id}/restore/": {
      "post": {
        "operationId": "clients.client.restore",
        "summary": "Restore Client",
        "description": "Return a retired partner to current use, clearing the retirement state.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "clients.client.restore",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/document-templates/signatories/{signatory_id}/status/": {
      "post": {
        "operationId": "document_templates.signatory.change_status",
        "summary": "Change Signatory Status",
        "description": "Move a signatory between draft, active, and inactive. Nothing is deleted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "document_templates"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "document_templates.signatory.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "signatory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/document-templates/templates/{template_id}/status/": {
      "post": {
        "operationId": "document_templates.template.change_status",
        "summary": "Change Document Template Status",
        "description": "Move a template between draft, active, and inactive. Existing documents are unaffected.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "document_templates"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "document_templates.template.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/archive/": {
      "post": {
        "operationId": "documents.document.archive",
        "summary": "Archive Document",
        "description": "Retire a document from active work with a mandatory reason. Nothing is deleted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "documents.document.archive",
        "x-risk-level": "high",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/restore/": {
      "post": {
        "operationId": "documents.document.restore",
        "summary": "Restore Document",
        "description": "Return an archived document to draft so it can be edited again.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "documents.document.restore",
        "x-risk-level": "high",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/documents/{document_id}/status/": {
      "post": {
        "operationId": "documents.document.change_status",
        "summary": "Change Document Status",
        "description": "Move a document between draft and ready. Cannot set archived.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "documents.document.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/files/{file_id}/archive/": {
      "post": {
        "operationId": "uploaded_files.file.archive",
        "summary": "Archive File",
        "description": "Retire a file from active work. A reason is required. Nothing is deleted and the version chain is unaffected. Admin authority only.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "uploaded_files"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "uploaded_files.file.archive",
        "x-risk-level": "high",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/files/{file_id}/restore/": {
      "post": {
        "operationId": "uploaded_files.file.restore",
        "summary": "Restore File",
        "description": "Return an archived file to active work. Admin authority only.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "uploaded_files"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "uploaded_files.file.restore",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/offers/conditions/{condition_id}/status/": {
      "post": {
        "operationId": "offers.condition.change_status",
        "summary": "Change Offer Condition Status",
        "description": "Mark a condition satisfied, waived, not applicable, or pending again.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "offers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "offers.condition.change_status",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "Resource payload — see the owning app's DATA_CONTRACT.md."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "APP_RESOURCE_REASON code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "previous": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}